# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Acortador de URLs",
        "description": "test",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/UrlShortener/shorten": {
            "post": {
                "tags": [
                    "UrlShortener"
                ],
                "summary": "Acortar una URL",
                "operationId": "8e8abc42472d364d477afcfb2c4cb306",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "original_url": {
                                        "type": "string",
                                        "example": "https://example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "URL acortada exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/v1/UrlShortener/{shortened}": {
            "get": {
                "tags": [
                    "UrlShortener"
                ],
                "summary": "Redireccionar a la URL original",
                "operationId": "ee11622be0ba5419d10cd57db886ebb2",
                "parameters": [
                    {
                        "name": "shortened",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redireccionando a la URL original"
                    },
                    "404": {
                        "description": "URL no encontrada"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "UrlShortener",
            "description": "UrlShortener"
        }
    ]
}